+++ Changelog.md
@@ -6,6 +6,7 @@ OpenCore Changelog
 - Disabled not present DVL0 device in SSDT-SBUS-MCHC by default, thx @stevezhengshiqi
 - Added Unicode font pages for some accented characters, plus box drawing and block elements, to `Builtin` text renderer
 - Improved support for overlong menu entries and very narrow console modes in builtin picker
+- Made `Builtin` text renderer ignore UI Scale, when required to ensure that text mode reaches minimum supported size of 80x25
 
 #### v0.9.2
 - Added `DisableIoMapperMapping` quirk, thx @CaseySJ